requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
attrs==25.1.0
audioop-lts==0.2.1
discord.py==2.5.0
fire==0.7.0
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
propcache==0.3.0
setuptools==75.8.2
termcolor==2.5.0
yarl==1.18.3
